editing some supporting documents
I've added a README for the markers to get a general overview of the
project.

I've made minor updates to the mini-milestones, adding a suggestion on
how to indicate whether a task has been completed or not.

I changed the name of the README.md to GIT_README because the original
readme was for the git, and now we have a general README for the
development branch / the project itself.